add flag to force_api_key, defaults to False
when api_key is forced the connection to stash will persist past cookie expiration this is useful for long running connections to stash that need to last longer than normal

this may cause issues if stash is expecting a cookie from a plugin, especially with hooks, note in the stash manual

"It is possible for hooks to trigger each other or themselves if they perform mutations. For safety, hooks will not be triggered if they have already been triggered in the context of the operation. Stash uses cookies to track this context, so it's important for plugins to send cookies when performing operations."